refactor: isolate admin/books components
