+++ src/app/admin/books/book-item.tsx
@@ -2,12 +2,12 @@
 import { type Book } from "@prisma/client";
 
 import { cn } from "@/lib/utils";
+import { buttonVariants } from "@/components/ui/button";
 
-import { buttonVariants } from "../ui/button";
-import { DeleteButton } from "./delete-button";
-import { EditBook } from "./edit";
+import { EditBook } from "./edit-book";
+import { DeleteBook } from "./delete-book";
 
-export const WorkItem = ({ book }: { book: Book }) => {
+export const BookItem = ({ book }: { book: Book }) => {
   return (
     <div className="border rounded-lg w-full p-3">
       {/* info */}
@@ -22,8 +22,7 @@ export const WorkItem = ({ book }: { book: Book }) => {
       <div className="flex items-center justify-between">
         <div className="space-x-3">
           <EditBook book={book} />
-
-          <DeleteButton book={book} />
+          <DeleteBook book={book} />
         </div>
         <a
           href={book.fileUrl}
