+++ src/app/admin/books/edit-book.tsx
@@ -1,7 +1,12 @@
 "use client";
 import { useState } from "react";
 import { useRouter } from "next/navigation";
+import type { Book, Language } from "@prisma/client";
 
+import { Button } from "@/components/ui/button";
+import { Input } from "@/components/ui/input";
+import { Label } from "@/components/ui/label";
+import { useToast } from "@/components/ui/use-toast";
 import {
   Dialog,
   DialogContent,
@@ -9,44 +14,45 @@ import {
   DialogTitle,
   DialogTrigger,
 } from "@/components/ui/dialog";
-import { Button, buttonVariants } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import { Upload } from "@/components/upload";
-import { useToast } from "@/components/ui/use-toast";
-import { Icons } from "../icons";
 
-export const NewBook = () => {
+interface EditBookFormValues {
+  title: string;
+  grade: string;
+  language: Language;
+  term: string;
+}
+
+export const EditBook = ({ book }: { book: Book }) => {
   const router = useRouter();
+
   const { toast } = useToast();
 
   const grades = ["7", "8", "9", "10", "11", "12"];
   const languages = ["T1", "T2"];
   const terms = ["1", "2", "3", "4"];
 
-  const [data, setData] = useState({
-    title: "",
-    fileUrl: "",
-    grade: grades[0],
-    language: languages[0],
-    term: terms[0],
+  const [data, setData] = useState<EditBookFormValues>({
+    title: book.title,
+    grade: book.grade,
+    language: book.language,
+    term: book.term,
   });
 
   const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
     event.preventDefault();
 
     // prisma call
-    const res = await fetch("/api/add-book", {
+    const res = await fetch("/api/edit-book", {
       method: "POST",
-      body: JSON.stringify(data),
+      body: JSON.stringify({ id: book.id, data }),
     });
 
     if (res.ok) {
       const data = await res.json();
 
       toast({
-        title: "Шығарма сәтті жүктелді",
-        description: `Жүктелген шығарма: ${data.book.title}`,
+        title: "Шығарма сәтті өзгертілді",
+        description: `Өзгертілген шығарма: ${data.book.title}`,
       });
       return router.push("/admin/books");
     }
@@ -60,14 +66,11 @@ export const NewBook = () => {
   return (
     <Dialog>
       <DialogTrigger asChild>
-        <Button className="bg-[#6C63FF] hover:bg-[#6C63FF]/90 text-sm gap-2">
-          <Icons.plus width={20} height={20} />
-          Жаңа шығарма
-        </Button>
+        <Button size="sm">Өзгерту</Button>
       </DialogTrigger>
       <DialogContent className="sm:max-w-[425px]">
         <DialogHeader>
-          <DialogTitle>Жаңа шығарма</DialogTitle>
+          <DialogTitle>Шығарманы өзгерту</DialogTitle>
         </DialogHeader>
         <form onSubmit={handleSubmit} className="space-y-4">
           <div>
@@ -75,32 +78,17 @@ export const NewBook = () => {
             <Input
               id="title"
               type="text"
+              value={data.title || ""}
               placeholder="Абай жолы"
               onChange={(e) => setData({ ...data, title: e.target.value })}
             />
-          </div>
-          <div>
-            <Upload data={data} setData={setData} />
-            {data.fileUrl && (
-              <div className="flex items-center space-x-3 my-3">
-                <a
-                  href={data.fileUrl}
-                  target="_blank"
-                  rel="noreferrer"
-                  className={buttonVariants({ variant: "outline" })}
-                >
-                  Көру
-                </a>
-                <span>Файл жүктелді</span>
-              </div>
-            )}
           </div>
           <div>
             <Label htmlFor="grade">Сынып</Label>
             <select
               id="grade"
               className="w-full px-3 py-2 rounded-md bg-transparent border"
-              defaultValue={grades[0]}
+              defaultValue={data.grade || grades[0]}
               onChange={(e) => setData({ ...data, grade: e.target.value })}
             >
               {grades.map((grade) => (
@@ -111,12 +99,14 @@ export const NewBook = () => {
             </select>
           </div>
           <div>
-            <label htmlFor="language">Тіл</label>
+            <Label htmlFor="language">Тіл</Label>
             <select
               id="language"
               className="w-full px-3 py-2 rounded-md bg-transparent border"
-              defaultValue="T1"
-              onChange={(e) => setData({ ...data, language: e.target.value })}
+              defaultValue={data.language || languages[0]}
+              onChange={(e) =>
+                setData({ ...data, language: e.target.value as Language })
+              }
             >
               {languages.map((language) => (
                 <option key={language} value={language}>
@@ -130,7 +120,7 @@ export const NewBook = () => {
             <select
               id="term"
               className="w-full px-3 py-2 rounded-md bg-transparent border"
-              defaultValue="1"
+              defaultValue={data.term || terms[0]}
               onChange={(e) => setData({ ...data, term: e.target.value })}
             >
               {terms.map((term) => (
