+++ src/app/admin/books/new-book.tsx
@@ -1,58 +1,52 @@
 "use client";
 import { useState } from "react";
 import { useRouter } from "next/navigation";
-import type { Book, Language } from "@prisma/client";
 
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import { useToast } from "@/components/ui/use-toast";
 import {
   Dialog,
   DialogContent,
   DialogHeader,
   DialogTitle,
   DialogTrigger,
-} from "../ui/dialog";
+} from "@/components/ui/dialog";
+import { Button, buttonVariants } from "@/components/ui/button";
+import { Input } from "@/components/ui/input";
+import { Label } from "@/components/ui/label";
+import { Upload } from "@/components/upload";
+import { useToast } from "@/components/ui/use-toast";
+import { Icons } from "../../../components/icons";
 
-interface EditBookFormValues {
-  title: string;
-  grade: string;
-  language: Language;
-  term: string;
-}
-
-export const EditBook = ({ book }: { book: Book }) => {
+export const NewBook = () => {
   const router = useRouter();
-
   const { toast } = useToast();
 
   const grades = ["7", "8", "9", "10", "11", "12"];
   const languages = ["T1", "T2"];
   const terms = ["1", "2", "3", "4"];
 
-  const [data, setData] = useState<EditBookFormValues>({
-    title: book.title,
-    grade: book.grade,
-    language: book.language,
-    term: book.term,
+  const [data, setData] = useState({
+    title: "",
+    fileUrl: "",
+    grade: grades[0],
+    language: languages[0],
+    term: terms[0],
   });
 
   const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
     event.preventDefault();
 
     // prisma call
-    const res = await fetch("/api/edit-book", {
+    const res = await fetch("/api/add-book", {
       method: "POST",
-      body: JSON.stringify({ id: book.id, data }),
+      body: JSON.stringify(data),
     });
 
     if (res.ok) {
       const data = await res.json();
 
       toast({
-        title: "Шығарма сәтті өзгертілді",
-        description: `Өзгертілген шығарма: ${data.book.title}`,
+        title: "Шығарма сәтті жүктелді",
+        description: `Жүктелген шығарма: ${data.book.title}`,
       });
       return router.push("/admin/books");
     }
@@ -66,11 +60,14 @@ export const EditBook = ({ book }: { book: Book }) => {
   return (
     <Dialog>
       <DialogTrigger asChild>
-        <Button size="sm">Өзгерту</Button>
+        <Button className="bg-[#6C63FF] hover:bg-[#6C63FF]/90 text-sm gap-2">
+          <Icons.plus width={20} height={20} />
+          Жаңа шығарма
+        </Button>
       </DialogTrigger>
       <DialogContent className="sm:max-w-[425px]">
         <DialogHeader>
-          <DialogTitle>Шығарманы өзгерту</DialogTitle>
+          <DialogTitle>Жаңа шығарма</DialogTitle>
         </DialogHeader>
         <form onSubmit={handleSubmit} className="space-y-4">
           <div>
@@ -78,17 +75,32 @@ export const EditBook = ({ book }: { book: Book }) => {
             <Input
               id="title"
               type="text"
-              value={data.title || ""}
               placeholder="Абай жолы"
               onChange={(e) => setData({ ...data, title: e.target.value })}
             />
+          </div>
+          <div>
+            <Upload data={data} setData={setData} />
+            {data.fileUrl && (
+              <div className="flex items-center space-x-3 my-3">
+                <a
+                  href={data.fileUrl}
+                  target="_blank"
+                  rel="noreferrer"
+                  className={buttonVariants({ variant: "outline" })}
+                >
+                  Көру
+                </a>
+                <span>Файл жүктелді</span>
+              </div>
+            )}
           </div>
           <div>
             <Label htmlFor="grade">Сынып</Label>
             <select
               id="grade"
               className="w-full px-3 py-2 rounded-md bg-transparent border"
-              defaultValue={data.grade || grades[0]}
+              defaultValue={grades[0]}
               onChange={(e) => setData({ ...data, grade: e.target.value })}
             >
               {grades.map((grade) => (
@@ -99,14 +111,12 @@ export const EditBook = ({ book }: { book: Book }) => {
             </select>
           </div>
           <div>
-            <Label htmlFor="language">Тіл</Label>
+            <label htmlFor="language">Тіл</label>
             <select
               id="language"
               className="w-full px-3 py-2 rounded-md bg-transparent border"
-              defaultValue={data.language || languages[0]}
-              onChange={(e) =>
-                setData({ ...data, language: e.target.value as Language })
-              }
+              defaultValue="T1"
+              onChange={(e) => setData({ ...data, language: e.target.value })}
             >
               {languages.map((language) => (
                 <option key={language} value={language}>
@@ -120,7 +130,7 @@ export const EditBook = ({ book }: { book: Book }) => {
             <select
               id="term"
               className="w-full px-3 py-2 rounded-md bg-transparent border"
-              defaultValue={data.term || terms[0]}
+              defaultValue="1"
               onChange={(e) => setData({ ...data, term: e.target.value })}
             >
               {terms.map((term) => (
